+++ src/Components/FardosaGedi/index.css
@@ -1,69 +1,57 @@
-/* App.css */
-/*body {
-    font-family: Arial, sans-serif;
-    background-color: #f4f4f9;
-    margin: 0;
-    padding: 20px;
-    color: #333;
-  }*/
-  
-/*.App {
-    max-width: 600px;
-    margin: auto;
-    background: white;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0,0,0,0.1);
-  }*/
-  
-  h2 {
-    color: #4A90E2;
-  }
-  
-  form {
-    display: flex;
-    flex-direction: column;
-    width: 600px;
-  }
+.formContainer {
+  width: 400px;
+  margin: 0 auto;
+  padding: 20px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  margin-top: 2px;
+}
 
-  .formContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    
-  }
+.formContainer h2 {
+  font-size: 1.5em;
+  color: rgb(6, 188, 243);
+
+}
+
+#time{
+  width: 100%;
+  padding: 4px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  margin-bottom: 10px;
+}
+
+.formContainer label {
+  display: block;
+  margin-bottom: 10;
+}
+
+.formContainer input[type="text"] {
+  width: 100%;
+  padding: 4px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+}
+
+.formContainer input[type="date"] {
+  width: 100%;
+  padding: 4px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+}
+
+.formContainer button[type="submit"] {
+  display: block;
+  width: 100%;
+  padding: 4px;
   
-  .formContainer form {
-    width: 80%; 
-  }
-  
-  
-  input[type="text"] {
-    padding: 8px;
-    margin-top: 5px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-  }
-  
-  button {
-    cursor: pointer;
-    background-color:#4A90E2;
-    color: white;
-    border: none;
-    padding: 10px 15px;
-    margin-top: 10px;
-    border-radius: 5px;
-    font-size: 16px;
-  }
-  
-  button:hover {
-    background-color: black;
-  }
-  
-  /* Responsive adjustments */
-  @media (max-width: 768px) {
-    .App {
-      padding: 15px;
-    }
-  }
-  
+  background-color: #007bff;
+  color: #fff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.formContainer button[type="submit"]:hover {
+  background-color: #0056b3;
+}
