+++ src/Components/Header/Header.js
@@ -78,7 +78,7 @@ export default function Header() {
         <ul>
           <li><Link to="/">Home</Link></li>
           <li><Link to="/about">About</Link></li>
-          <li><a href="#">Jobs</a></li>
+          <li><Link to="/jobs">Jobs</Link></li>
         </ul>
       </div>
