+++ src/Components/Ignatius-Caesar/JobList.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect } from 'react'; 
 import './index.css'; // Ensure you have this CSS file in the same directory
 import JobTab from '../Stephen/JobTab';
